Add 'clear all completed' function
