+++ src/index.html
@@ -12,7 +12,7 @@
   <div class="container">
     <h2><span>Today's To Do</span><span class="material-symbols-outlined" >sync</span></h2>
     <form action="" id="form">
-      <input type="text" id="description" name="description" placeholder="Add to your list..." maxlength="30" />
+      <input type="text" id="description" name="description" placeholder="Add to your list..." autofocus />
       <button type="submit"><i class="material-symbols-outlined submit-arrow">reply</i></button>
     </form>
 
@@ -21,7 +21,7 @@
 
     </ul>
     
-    <button type="button" class="clear">Clear all completed</button>
+    <button type="button" class="clear" id="clear">Clear all completed</button>
   </div>
   
 </body>
